+++ src/export-button/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { ExportButton } from "./export-button";
+export { generateReport } from "./utils/export-utils";