+++ src/table/index.js
@@ -1,8 +1,8 @@
 // @flow
-import { type PagedTableProps } from "./paged-table";
+import type { PagedTableProps } from "./paged-table";
 
 export { PagedTable } from "./paged-table";
 export { SimpleTable } from "./simple-table";
 export { LargeSimpleTable } from "./large-simple-table";
-export { ExportTableButton } from "./components/export-button";
+export { ExportTableButton } from "./components/export-table-button";
 export type { PagedTableProps };